Split rooms in different files
