+++ jupyter_collaboration/rooms/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) Jupyter Development Team.
+# Distributed under the terms of the Modified BSD License.
+
+from .document import DocumentRoom  # noqa
+from .transient import TransientRoom  # noqa
